requirements.txt
asgiref==3.5.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
constantly==15.1.0
cryptography==37.0.1
daphne==3.0.2
debugpy @ file:///D:/bld/debugpy_1649586532851/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.0.4
djangorestframework==3.13.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel @ file:///D:/bld/ipykernel_1649684466269/work/dist/ipykernel-6.13.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1648413712410/work
jedi @ file:///D:/bld/jedi_1649067336192/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1650972844302/work
jupyter-core @ file:///D:/bld/jupyter_core_1645024565859/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
psutil @ file:///D:/bld/psutil_1648873130737/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyOpenSSL==22.0.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2020.1
pywin32==303
pyzmq @ file:///D:/bld/pyzmq_1649055863315/work
service-identity==21.1.0
short-url==1.2.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.3.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
tornado @ file:///D:/bld/tornado_1648827438424/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
wincertstore==0.2
zope.interface==5.4.0
